implement a way to provide missing motor parameters
The CANOpen objects that store the factors are not saved to
non-volatile memory. They are therefore basically useless
for autoconfiguration

One can set the factors explicitely, and still use the rest
of the API as-if they were coming from the device. Only the
rated current can be used. Given how critical it is (the
drive reports currents as thousands of the rated current),
this one is left out and must be queried beforehand. The
MotorParameters structure has instead the torque constant